dependency injection in prep for tests
